+++ packages/framer-motion/src/gestures/__tests__/pan.test.tsx
@@ -20,6 +20,7 @@ describe("pan", () => {
                         onPanStart={() => {
                             count += increment
                             setIncrement(2)
+                            console.log("on pan start")
                         }}
                         onPan={() => (count += increment)}
                         onPanEnd={() => {